extracting views into their own files
